+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataexchange/aggregate/Filter.java
@@ -25,10 +25,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.dataexchange.analytics;
+package org.hisp.dhis.dataexchange.aggregate;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.List;
 
 import lombok.Getter;
@@ -42,24 +41,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @Setter
 @NoArgsConstructor
 @Accessors( chain = true )
-public class SourceRequest
+public class Filter
     implements Serializable
 {
     @JsonProperty
-    private List<String> dx = new ArrayList<>();
+    private String dimension;
 
     @JsonProperty
-    private List<String> pe = new ArrayList<>();
-
-    @JsonProperty
-    private List<String> ou = new ArrayList<>();
-
-    @JsonProperty
-    private List<Filter> filters = new ArrayList<>();
-
-    @JsonProperty
-    private String inputIdScheme;
-
-    @JsonProperty
-    private String outputIdScheme;
+    private List<String> items;
 }
