+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataexchange/aggregate/TargetType.java
@@ -25,11 +25,10 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.dataexchange.analytics;
+package org.hisp.dhis.dataexchange.aggregate;
 
-import org.hisp.dhis.common.IdentifiableObjectStore;
-
-public interface AggregateDataExchangeStore
-    extends IdentifiableObjectStore<AggregateDataExchange>
+public enum TargetType
 {
+    INTERNAL,
+    EXTERNAL;
 }
