+++ dhis-2/dhis-services/dhis-service-data-exchange/src/main/java/org/hisp/dhis/dataexchange/aggregate/AggregateDataExchangeService.java
@@ -44,13 +44,6 @@ import org.hisp.dhis.analytics.DataQueryService;
 import org.hisp.dhis.common.DimensionalObject;
 import org.hisp.dhis.common.Grid;
 import org.hisp.dhis.common.IdScheme;
-import org.hisp.dhis.dataexchange.analytics.AggregateDataExchange;
-import org.hisp.dhis.dataexchange.analytics.AggregateDataExchangeStore;
-import org.hisp.dhis.dataexchange.analytics.Api;
-import org.hisp.dhis.dataexchange.analytics.Filter;
-import org.hisp.dhis.dataexchange.analytics.SourceRequest;
-import org.hisp.dhis.dataexchange.analytics.TargetRequest;
-import org.hisp.dhis.dataexchange.analytics.TargetType;
 import org.hisp.dhis.dataexchange.client.Dhis2Client;
 import org.hisp.dhis.dataexchange.client.Dhis2Config;
 import org.hisp.dhis.dxf2.common.ImportOptions;
