+++ dhis-2/dhis-services/dhis-service-data-exchange/src/test/java/org/hisp/dhis/dataexchange/aggregate/AggregateDataExchangeServiceTest.java
@@ -45,14 +45,6 @@ import org.hisp.dhis.common.BaseDimensionalObject;
 import org.hisp.dhis.common.DimensionType;
 import org.hisp.dhis.common.DimensionalObject;
 import org.hisp.dhis.common.IdScheme;
-import org.hisp.dhis.dataexchange.analytics.AggregateDataExchange;
-import org.hisp.dhis.dataexchange.analytics.AggregateDataExchangeStore;
-import org.hisp.dhis.dataexchange.analytics.Api;
-import org.hisp.dhis.dataexchange.analytics.Source;
-import org.hisp.dhis.dataexchange.analytics.SourceRequest;
-import org.hisp.dhis.dataexchange.analytics.Target;
-import org.hisp.dhis.dataexchange.analytics.TargetRequest;
-import org.hisp.dhis.dataexchange.analytics.TargetType;
 import org.hisp.dhis.dataexchange.client.Dhis2Client;
 import org.hisp.dhis.dxf2.common.ImportOptions;
 import org.hisp.dhis.dxf2.datavalueset.DataValueSet;
